+++ controlplane/CHANGELOG.md
@@ -4,6 +4,35 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.37.2](https://github.com/wundergraph/cosmo/compare/controlplane@0.37.1...controlplane@0.37.2) (2023-11-20)
+
+### Bug Fixes
+
+* move to bitnami charts and exit 1 on migration issues ([#275](https://github.com/wundergraph/cosmo/issues/275)) ([90d9d93](https://github.com/wundergraph/cosmo/commit/90d9d938cefdc78a9f34d69387f306b4d691c7f0)) (@StarpTech)
+* remove unnecessary fields and add populate the createdBy for compositions ([#272](https://github.com/wundergraph/cosmo/issues/272)) ([82b716c](https://github.com/wundergraph/cosmo/commit/82b716cb629e4c84e5cf45461951594abab9df6b)) (@JivusAyrus)
+
+## [0.37.1](https://github.com/wundergraph/cosmo/compare/controlplane@0.37.0...controlplane@0.37.1) (2023-11-17)
+
+### Bug Fixes
+
+* show latest valid subgraph schema ([#259](https://github.com/wundergraph/cosmo/issues/259)) ([d954b91](https://github.com/wundergraph/cosmo/commit/d954b91bd212ae1a33257c662a4ff8a2ac8c2b56)) (@JivusAyrus)
+
+# [0.37.0](https://github.com/wundergraph/cosmo/compare/controlplane@0.36.0...controlplane@0.37.0) (2023-11-15)
+
+### Features
+
+* consider input and argument usage for breaking change detection ([#255](https://github.com/wundergraph/cosmo/issues/255)) ([e10ac40](https://github.com/wundergraph/cosmo/commit/e10ac401f543f5540b5ada8f80533ddfbd0bc728)) (@jensneuse)
+
+# [0.36.0](https://github.com/wundergraph/cosmo/compare/controlplane@0.35.0...controlplane@0.36.0) (2023-11-15)
+
+### Bug Fixes
+
+* remove 10 day limit for free trial ([#260](https://github.com/wundergraph/cosmo/issues/260)) ([9e59f58](https://github.com/wundergraph/cosmo/commit/9e59f583cd7195a74012795e3e2401ae9cae4bfb)) (@JivusAyrus)
+
+### Features
+
+* add check for deleted subgraphs ([#258](https://github.com/wundergraph/cosmo/issues/258)) ([ba87fe5](https://github.com/wundergraph/cosmo/commit/ba87fe51631ece9c2efaea6350dc93590f1846c5)) (@Pagebakers)
+
 # [0.35.0](https://github.com/wundergraph/cosmo/compare/controlplane@0.34.0...controlplane@0.35.0) (2023-11-10)
 
 ### Features